Added support basis for `msys` and `cygwin` terminal emulators.
To create the basic structure with which to support the two terminal
emulators listed above, the employment of a DOS command was necessary.
In order to facilitate this BASH was parsed the command "cmd /c", which
permits the parsing of a DOS command from the BASH shell to a DOS-like
shell as a string-type argument.
An example of this is as follows: `cmd /c "echo Hello, World!"`.

The command in DOS (and Windows environments by extension) for the
creation of a symbolic link on the command line is `mklink`. Said
command does not permit the user to create symbolic links to directores
without the employment of the `/d` flag. As such specific
considerations needed to be made for this I have opted to rely on the
unecessarily complex option of appending a forward-slash ('/') to any
links requiring treatment as a directory. In order to ensure that the
directory flag was specified a REGEX pattern is employed as a pattern
with grep to search for a forward-slash character preceeding one or more
normal "word" characters (that being any letter) within the string.
Grep will return an integer specifying the amount of matches, this being
from 0 onwards. If this returned value is greater than 0 then the
directory flag will be employed with the `mklink` DOS command.

This addition should not (to the extents of my knowledge) result in
adverse effect of GNU-Linux or similar *nix-like environments thanks to
its implementation. However this script was not made with POSIX
compliance in mind and as such the invocation of this script under such
environments must be done with extreme prejudice.